requirements.txt
pytorch==1.13.1
torchvision==0.14.1
torchaudio==0.13.1
opencv-python-headless==4.5.5.0  # Ensure compatibility with OpenCV 4
numpy==1.21.5 
matplotlib==3.5.1
glob2==0.7
tqdm==4.6.0 
pathlib==1.0.1 
argparse==1.6.1
piqa
tqdm